split interface more
