+++ cl/phase1/core/transition/machine/machine.go
@@ -3,7 +3,9 @@ package machine
 
 import (
 	"github.com/ledgerwatch/erigon/cl/cltypes"
+	"github.com/ledgerwatch/erigon/cl/cltypes/solid"
 	"github.com/ledgerwatch/erigon/cl/phase1/core/state"
+	"github.com/ledgerwatch/erigon/core/types"
 )
 
 type Interface interface {
@@ -12,33 +14,35 @@ type Interface interface {
 	SlotProcessor
 }
 
+type BlockProcessor interface {
+	BlockHeaderProcessor
+	BlockOperationProcessor
+}
+
 type BlockValidator interface {
 	VerifyBlockSignature(s *state.BeaconState, block *cltypes.SignedBeaconBlock) error
 	VerifyTransition(s *state.BeaconState, block *cltypes.BeaconBlock) error
 }
 
-// TransitionState will call impl..ProcessSlots, then impl.VerifyBlockSignature, then ProcessBlock, then impl.VerifyTransition
-func TransitionState(impl Interface, s *state.BeaconState, block *cltypes.SignedBeaconBlock) error {
-	currentBlock := block.Block
-	if err := impl.ProcessSlots(s, currentBlock.Slot); err != nil {
-		return err
-	}
+type SlotProcessor interface {
+	ProcessSlots(s *state.BeaconState, slot uint64) error
+}
 
-	if err := impl.VerifyBlockSignature(s, block); err != nil {
-		return err
-	}
+type BlockHeaderProcessor interface {
+	ProcessBlockHeader(s *state.BeaconState, block *cltypes.BeaconBlock) error
+	ProcessWithdrawals(s *state.BeaconState, withdrawals *solid.ListSSZ[*types.Withdrawal]) error
+	ProcessExecutionPayload(s *state.BeaconState, payload *cltypes.Eth1Block) error
+	ProcessRandao(s *state.BeaconState, randao [96]byte, proposerIndex uint64) error
+	ProcessEth1Data(state *state.BeaconState, eth1Data *cltypes.Eth1Data) error
+	ProcessSyncAggregate(s *state.BeaconState, sync *cltypes.SyncAggregate) error
+	VerifyKzgCommitmentsAgainstTransactions(transactions *solid.TransactionsSSZ, kzgCommitments *solid.ListSSZ[*cltypes.KZGCommitment]) (bool, error)
+}
 
-	// Transition block
-	if err := ProcessBlock(impl, s, block); err != nil {
-		return err
-	}
-
-	// perform validation
-	if err := impl.VerifyTransition(s, currentBlock); err != nil {
-		return err
-	}
-
-	// if validation is successful, transition
-	s.SetPreviousStateRoot(currentBlock.StateRoot)
-	return nil
+type BlockOperationProcessor interface {
+	ProcessProposerSlashing(s *state.BeaconState, propSlashing *cltypes.ProposerSlashing) error
+	ProcessAttesterSlashing(s *state.BeaconState, attSlashing *cltypes.AttesterSlashing) error
+	ProcessAttestations(s *state.BeaconState, attestations *solid.ListSSZ[*solid.Attestation]) error
+	ProcessDeposit(s *state.BeaconState, deposit *cltypes.Deposit) error
+	ProcessVoluntaryExit(s *state.BeaconState, signedVoluntaryExit *cltypes.SignedVoluntaryExit) error
+	ProcessBlsToExecutionChange(state *state.BeaconState, signedChange *cltypes.SignedBLSToExecutionChange) error
 }
